feat(midi): clip-file note pairing, path I/O, and format sniffing
The library half of #50. A clip-file importer needed three things the
engine did not offer, so each caller would have had to build them:

- ParsedClipFile::notes() pairs the event stream into whole notes with
  durations, keyed on (group, channel, note) so a Note Off only closes a
  Note On in its own group and channel. Velocity stays 16-bit: pairing
  here is what stops a caller reaching for velocity_u7 and narrowing a
  clip file's full-resolution velocity on the way in. The MIDI-2
  counterpart of smf::tracks, whose per-channel pairing this mirrors.

- tutti_midi_io::clip gives clip files the path-level codec SMF already
  had (read_clip_file_from_path / write_clip_file_to_path), so callers
  stop hand-rolling fs::read plus error mapping.

- MidiFileKind::sniff{,_path} decides SMF vs clip file by magic instead
  of by extension. Extension cannot do this job: .mid and .midi are used
  for both formats in the wild, and M2-116 §5 adds .midi2 without
  retiring them. sniff_path reads 8 bytes, and separates "unreadable"
  (Err) from "readable but neither format" (Ok(None)) so a caller can
  fall through to another importer without swallowing real IO errors.

Tests cover the pairing rules that are easy to get wrong: cross-channel
and cross-group isolation, LIFO close of overlapping identical notes,
MIDI 1.0 velocity-0 Note On as a Note Off (built via from_midi1_bytes,
since note_on_7bit widens to CV2 and would never reach that path), notes
left open at End of Clip being dropped, and a clip file misnamed .mid
being identified correctly anyway.

Also clears three clippy warnings in tests added earlier in this branch.

App wiring stays out of scope — it is blocked on the projection rebuild.

+++ crates/midi/tutti-midi-types/src/ump/stream.rs
@@ -354,7 +354,7 @@ mod tests {
         assert_eq!(out.len(), 1);
         match UmpMessage::try_from(out[0].data_words()).unwrap() {
             UmpMessage::UmpStream(UmpStream::FunctionBlockName(m)) => {
-                assert_eq!(u8::from(m.function_block()), 2);
+                assert_eq!(m.function_block(), 2);
                 assert_eq!(m.name(), "Keys");
             }
             other => panic!("expected FunctionBlockName, got {other:?}"),
@@ -369,7 +369,7 @@ mod tests {
         let words: Vec<u32> = out.iter().flat_map(|e| e.data_words().to_vec()).collect();
         match UmpMessage::try_from(&words[..]).unwrap() {
             UmpMessage::UmpStream(UmpStream::FunctionBlockName(m)) => {
-                assert_eq!(u8::from(m.function_block()), 7);
+                assert_eq!(m.function_block(), 7);
                 assert_eq!(m.name(), long);
             }
             other => panic!("expected FunctionBlockName, got {other:?}"),
